requirements.txt
acvl-utils>=0.2
dynamic-network-architectures>=0.2
tqdm
dicom2nifti
scikit-image>=0.14
medpy
scipy
batchgenerators>=0.25
numpy
scikit-learn
scikit-image>=0.19.3
SimpleITK>=2.2.1
pandas
graphviz
tifffile
requests
nibabel
matplotlib
seaborn
imagecodecs
yacs
monai==1.3.0
opencv-python
torchinfo
einops
ninja
timm 
causal-conv1d
mamba-ssm
numba